M - Ya crea la BD Room

+++ KotlinApp/app/src/main/java/com/example/kotlinapp/data/room/entity/TrackpointEntity.kt
@@ -18,7 +18,7 @@ import com.example.kotlinapp.model.Ruta
     ]
 )
 data class TrackpointEntity(
-    val idRuta: Long,
+    val idRuta: Int,
     val posicion: Int,
     val latitud: Double,
     val longitud: Double,
